requirements.txt
Appium-Python-Client==1.1.0
jarowinkler==1.1.1
jieba==0.42.1
joblib==1.1.0
Levenshtein==0.18.1
numpy==1.21.6
opencv-python==4.5.5.62
rapidfuzz==2.1.2
scikit-learn==1.0.2
scipy==1.7.3
selenium==3.141.0
threadpoolctl==3.1.0
uiautomator==1.0.2
urllib3==1.26.9
